Fix editing patterns. Use the post_title instead of the post_name when generating the posts, as post_name can’t be predicted.
Use post_title instead of post_name, because post_name will start appending pattern-name-1 to posts if they already exist. That means that pattern-name-1 becomes pattern-name-1-1 if there’s already a garbage post in the database for that. post_title will always be what we set it to, so we should be able to deped on it.